+++ bad6.js
@@ -3,14 +3,18 @@ import { setupStats } from "./modules/listenerFunctions.js";
 import { registerHandlebarsHelpers } from "./modules/utils.js";
 import { preloadHandlebarsTemplates } from "./modules/utils.js"; 
 import { loadChartJS } from "./modules/objects/stat-chart-loader.js";
-import { BAD6Roller } from './modules/apps/bad6-roller.js';
+import { rollerControl } from './modules/apps/bad6-roller.js';
+import { HueShiftControl } from "./modules/apps/hue-shift.js";
+import { outroControl } from "./modules/apps/stylizedOutro.js";
 
 // Wait for Chart.js to load before initializing the system
 Hooks.once("init", async () => {
     console.log("BAD6 Core System is Initializing");
     // Load Chart.js
     await loadChartJS();
-    BAD6Roller();
+    rollerControl();
+    HueShiftControl();
+    outroControl();
     console.log("Chart.js has been loaded");
     // Dynamically import UserSheet and StandSheet after Chart.js is loaded
     const { UserSheet } = await import("./modules/sheets/user-actor-sheet.js");