+++ CHANGELOG.md
@@ -2,8 +2,14 @@
 
 ## [0.7.X] – 2025-06-29
 ### Added
-## BAD6 Roller
+## Imports
+Previously developed code which is now being added to the BAD6 System.
+# BAD6 Roller
 - Runs from Scene Controls (Token Controls)
+# Canvas Hue Shift
+- Runs from Scene Controls (Lighting Controls)
+# To Be Continued
+- Runs from Scene Controls (Lighting Controls)
 ### Optimization
 
 ### Fixes